Changed the default size for the NewPrinter window to 720x480
This is the minimum resolution used when in composite mode (NTSC), so it makes
sense to use it as reference so that we make sure that this window use at least
all the available space when running in those resolutions.

Note that this is not setting a minimum, so the window can be still resized down
if desired. Also, when launching the window in lower resolutions, the window
will simply launch maximized, so it should be all good.